+++ R/esr.R
@@ -8,16 +8,13 @@
 #'
 #' @references
 #' Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.
-#'
-#' @param x An MCMC object
-#' @param by A string indicating whether to return the estimates by
-#' the object ("all"), "parameter" or "term"
-#' @param as_df A flag indicating whether to return the estimates as a data frame versus a list.
-#' @param ... Unused
+#' @inheritParams rhat
 #' @return The esr value(s) as a data frame or list
 #' @export
 #' @examples
 #' esr(mcmcr_example)
+#' esr(mcmcrs(mcmcr_example, mcmcr_example))
+#' esr(mcmcrs(mcmcr_example, mcmcr_example), bound = TRUE)
 esr <- function(x, ...) {
   UseMethod("esr")
 }
@@ -70,6 +67,13 @@ esr.mcmcr <- function(x, by = "all", as_df = FALSE, ...) {
 
 #' @describeIn esr Effective Sampling Rate for an mcmcrs object
 #' @export
-esr.mcmcrs <- function(x, by = "all", ...) {
-  lapply(x, esr, by = by)
+esr.mcmcrs <- function(x, by = "all", as_df = FALSE, bound = FALSE, ...) {
+  check_flag(bound)
+  check_unused(...)
+
+  if(bound) {
+    x <- Reduce(bind_chains, x)
+    return(esr(x, by = by, as_df = as_df))
+  }
+  lapply(x, esr, by = by, as_df = as_df)
 }