+++ R/rhat.R
@@ -9,6 +9,7 @@
 #' "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the values as a
 #' data frame versus a named list.
+#' @param bound flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.
 #' @return The rhat value(s).
 #' @references
 #' Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
@@ -18,6 +19,8 @@
 #' rhat(mcmcr_example, by = "parameter")
 #' rhat(mcmcr_example, by = "term")
 #' rhat(mcmcr_example, by = "term", as_df = TRUE)
+#' rhat(mcmcrs(mcmcr_example, mcmcr_example))
+#' rhat(mcmcrs(mcmcr_example, mcmcr_example), bound = TRUE)
 rhat <- function(x, ...) {
   UseMethod("rhat")
 }
@@ -85,6 +88,13 @@ rhat.mcmcr <- function(x, by = "all", as_df = FALSE, ...) {
 
 #' @describeIn rhat R-hat for an mcmcrs object
 #' @export
-rhat.mcmcrs <- function(x, by = "all", as_df = FALSE, ...) {
+rhat.mcmcrs <- function(x, by = "all", as_df = FALSE, bound = FALSE, ...) {
+  check_flag(bound)
+  check_unused(...)
+
+  if(bound) {
+    x <- Reduce(bind_chains, x)
+    return(rhat(x, by = by, as_df = as_df))
+  }
   lapply(x, rhat, by = by, as_df = as_df)
 }